Migrating to control.Isabelle (instead of libisabelle). Status: compiles without errors but does not run

+++ src/main/scala/qrhl/tactic/WpTac.scala
@@ -1,16 +1,11 @@
 package qrhl.tactic
 
-import info.hupel.isabelle.Operation
-import info.hupel.isabelle.pure.{Term, Typ => ITyp}
-import qrhl.isabelle.{Isabelle, RichTerm}
-import qrhl.logic._
-import qrhl.{QRHLSubgoal, State, Subgoal, UserException}
+import qrhl.{QRHLSubgoal, State, Subgoal}
 
-import RichTerm.typ_tight_codec
-import RichTerm.term_tight_codec
+import isabelle.control.MLValue.Implicits._
 
 case class WpTac(left:Int, right:Int)
-  extends IsabelleTac[(BigInt,BigInt)]("wp_tac", { _ => (left,right) }) {
+  extends IsabelleTac[(Int,Int)]("wp_tac", { _ => (left,right) }) {
 
   assert(left >= 0, s"wp tactic must have nonnegative arguments (not left=$left)")
   assert(right >= 0, s"wp tactic must have nonnegative arguments (not right=$right)")
